refactor(create-recipe): move recipe serialization to a separate module
