+++ src/recipes/utils/serializing.ts
@@ -1,0 +1,10 @@
+export const serializeRecipe = (
+  name: string,
+  description: string,
+  ingredients: string[],
+) =>
+  JSON.stringify({
+    description,
+    ingredients: ingredients.map((ingredient) => ({ name: ingredient })),
+    name,
+  })
